backend/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
Pillow==8.3.1
flask-cors==3.0.10